Edit/remove groups. Add premissions block.

+++ server/src/contollers/users.js
@@ -1,17 +1,35 @@
 const UserDB = require("../db-models/User")
 const errorHandler = require('../utils/error-handler')
-const UserDto = require("../models/User");
 const bcrypt = require("bcryptjs");
+const User = require("../models/User");
+const UserDto = require("../models/UserDto");
 
 module.exports.getAllUsers = async (req, res) => {
     try {
+        const user = User.toModel(req.user)
+
+        if (!user.isAdmin) {
+            errorHandler(res, "Отказано в доступе!");
+            return;
+        }
+
         const users = await UserDB.find();
-        res.status(200).json(users.map((it) => UserDto.toDto(it)));
+        res.status(200).json(
+            users.map((it) => UserDto.toDto(it))
+                .sort((a, b) => a.name.localeCompare(b.name))
+        );
     } catch (e) {
         errorHandler(res, e);
     }
 }
 module.exports.createUser = async (req, res) => {
+    const user = User.toModel(req.user)
+
+    if (!user.isAdmin) {
+        errorHandler(res, "Отказано в доступе!");
+        return;
+    }
+
     try {
         if (!req.body.name) {
             res.status(404).json({
@@ -36,6 +54,13 @@ module.exports.createUser = async (req, res) => {
 }
 
 module.exports.updateUser = async (req, res) => {
+    const user = User.toModel(req.user)
+
+    if (!user.isAdmin) {
+        errorHandler(res, "Отказано в доступе!");
+        return;
+    }
+
     const updatedUser = {
         name: req.body.name,
         email: req.body.email,
@@ -68,6 +93,13 @@ module.exports.updateUser = async (req, res) => {
 }
 
 module.exports.removeUser = async (req, res) => {
+    const user = User.toModel(req.user)
+
+    if (!user.isAdmin) {
+        errorHandler(res, "Отказано в доступе!");
+        return;
+    }
+
     try {
         await UserDB.findByIdAndRemove(req.params.id);
         res.status(200).json();
